Move in_memory_proto_stream.* to internal.
